# qrng-client-api/openapi/qrng-public-v1.yaml
...
openapi: 3.0.3
info:
  title: Kapuã QRNG — API Pública
  version: 1.0.0
  description: >-
    API pública do gerador quântico de números aleatórios (QRNG) do Kapuã/Dobslit. Fonte física:
    FPGA Red Pitaya, stream uint32 little-endian, sem conditioning (confirmado no código-fonte do
    pipeline físico). Bytes brutos são servidos sem processamento adicional; a avaliação de
    min-entropia é feita separadamente pela suíte NIST SP 800-90B (ver aba Teste NIST / /qrng/nist/
    no frontend) — nenhum endpoint aqui certifica ou garante um nível de entropia.


    Unidade de TRANSPORTE (o que esta API entrega): bytes brutos, tal como lidos do registrador AXI
    FIFO da FPGA (4 bytes little-endian por amostra uint32). Unidade de AVALIAÇÃO ESTATÍSTICA (o que
    a suíte NIST SP 800-90B mede sobre uma amostra desses bytes): symbol width de 8 bits por
    decomposição byte a byte -- ver assessment_symbol_width nos resultados de /nist/*. As duas
    unidades são independentes: o formato de transporte não é uma alegação sobre quantos bits têm
    significado físico nem sobre origem quântica comprovada -- isso é o que a avaliação estatística
    separada existe para medir.
  contact:
    name: Dobslit
servers:
  - url: https://bongo.dobslit.com/qrng/v1
    description: Produção (via nginx, proxy para o client-api)
...
components:
  securitySchemes:
    bearerAuthJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Sessão de usuário (login via /v1/auth/login ou /v1/auth/register). Usada pelos endpoints
        /v1/auth/me, /v1/tokens e /v1/admin/*.
    bearerAuthToken:
      type: http
      scheme: bearer
      description: >-
        Token de API pessoal (formato dobslit_qrng_live_<hex>, emitido por POST /v1/tokens). Usado
        pelos endpoints de consumo: /v1/random, /v1/health. Endpoints /v1/me/* aceitam tanto este
        token quanto o JWT de sessão.
  schemas:
    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          example: req_a383397963db4222
          nullable: true
        error:
          type: string
          example: INVALID_TOKEN
        message:
          type: string
          example: Token inválido ou revogado.
      required:
        - error
    RandomResponse:
      type: object
      description: >-
        Corpo retornado quando format=hex|base64|uint8 OU quando o parâmetro format é omitido
        (default=hex). Para format=raw o corpo é application/octet-stream com exatamente os N bytes
        brutos (sem este JSON, sem prefixo, sem BOM) — ou use a rota dedicada GET /raw (e
        /public/raw).
      properties:
        request_id:
          type: string
          example: req_a383397963db4222
        source:
          type: string
          example: dobslit-qrng-ufpe-fpga
          description: Rótulo da fonte (env QRNG_SOURCE_LABEL).
        provenance:
          type: string
          enum:
            - live
            - replay
            - fixture
            - historical
            - fallback
            - unknown
          example: live
          description: >-
            Proveniência dos bytes. 'live' = fonte física em tempo real (produção). Um ambiente de
            staging que sirva fixtures/replay reporta 'replay'/'fixture' — nunca 'live'. Também no
            header X-QRNG-Provenance quando format=raw.
        bytes:
          type: integer
          example: 32
          description: Quantidade de bytes gerados.
        format:
          type: string
          enum:
            - hex
            - base64
            - uint8
          example: hex
        random:
          description: >-
            hex: string hexadecimal (2N caracteres). base64: string base64. uint8: array de N
            inteiros [0,255].
          oneOf:
            - type: string
              example: 5c0eda118fdf680778ecf97c4ab89cb7
            - type: array
              items:
                type: integer
                minimum: 0
                maximum: 255
        timestamp:
          type: string
          format: date-time
      required:
        - request_id
        - source
        - provenance
        - bytes
        - format
        - random
        - timestamp
...
  responses:
    Unauthorized:
      description: Token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Limite de requisições por IP ou por token excedido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Segundos até poder tentar novamente.
...
paths:
...
  /random:
    get:
      tags:
        - Random
      summary: Gera bytes aleatórios da fonte QRNG (FPGA Red Pitaya, uint32-LE, sem conditioning).
      description: >
        Confirmado no código-fonte de todo o pipeline físico (auditoria 2026-08-25): o upstream lê
        um registrador AXI FIFO diretamente via mmap na Red Pitaya e grava exatamente 4 bytes
        little-endian por amostra, sem nenhum processamento (whitening/debiasing/conditioning). Esta
        rota entrega esses bytes brutos -- nenhuma avaliação de entropia é feita aqui; ver a suíte
        NIST SP 800-90B separadamente para isso.
      security:
        - bearerAuthToken: []
      parameters:
        - name: bytes
          in: query
          schema:
            type: integer
            minimum: 1
            default: 32
          description: >-
            Quantidade de bytes a gerar. Limite por requisição = MAX_BYTES_PER_REQUEST (padrão 1
            MiB).
        - name: format
          in: query
          schema:
            type: string
            enum:
              - hex
              - base64
              - uint8
              - raw
          description: >
            hex (default — inclusive quando o parâmetro é OMITIDO): JSON RandomResponse com `random`
            = string hexadecimal de 2N caracteres. base64: JSON RandomResponse com `random` = string
            base64 dos N bytes. uint8: JSON RandomResponse com `random` = array de N inteiros
            [0,255]. raw: corpo application/octet-stream com EXATAMENTE os N bytes brutos — sem
            JSON, sem texto, sem prefixo, sem BOM; request_id e proveniência vão nos headers
            X-Request-Id / X-QRNG-Source / X-QRNG-Conditioned. COMPATIBILIDADE: omitir `format`
            retorna JSON (hex), NÃO binário — o binário exige `format=raw` explícito, ou a rota
            dedicada GET /raw.
      responses:
        '200':
          description: >
            Bytes gerados. format=hex|base64|uint8 (ou omitido) → application/json (RandomResponse).
            format=raw → application/octet-stream com N bytes exatos.
          headers:
            X-Request-Id:
              schema:
                type: string
              description: Identificador da requisição (também no corpo quando a resposta é JSON).
            X-QRNG-Source:
              schema:
                type: string
              description: 'Proveniência da fonte (ex.: dobslit-qrng-ufpe-fpga). Presente em format=raw.'
            X-QRNG-Conditioned:
              schema:
                type: string
                enum:
                  - 'false'
              description: Sempre 'false' — bytes não condicionados. Presente em format=raw.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomResponse'
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: bytes acima de MAX_BYTES_PER_REQUEST.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: bytes ou format inválidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          description: >
            Upstream retornou erro OU o Content-Type declarado não é suportado pelo contrato
            (UPSTREAM_UNSUPPORTED_CONTENT_TYPE, UPSTREAM_MISSING_CONTENT_TYPE,
            UPSTREAM_INVALID_JSON, UPSTREAM_JSON_SCHEMA_MISMATCH, UPSTREAM_LENGTH_MISMATCH,
            UPSTREAM_LEGACY_FORMAT_DISABLED) -- ver interpretUpstreamResponse().
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: >-
            Upstream indisponível, timeout, ou buffer com menos bytes que o pedido
            (INSUFFICIENT_ENTROPY).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        429_quota:
          description: (usa código 429) Cota diária de requisições ou bytes excedida.
  /raw:
    get:
      tags:
        - Random
      summary: >-
        Alias binário explícito de GET /random?format=raw — application/octet-stream com N bytes
        exatos.
      description: >
        Idêntico a GET /random?format=raw. Rota própria para consumidores que só querem os bytes
        crus sem lidar com o parâmetro format. NUNCA retorna JSON: o corpo são EXATAMENTE os N bytes
        (Content-Length = N), sem prefixo, texto ou BOM. request_id e proveniência vão nos headers
        X-Request-Id / X-QRNG-Source / X-QRNG-Conditioned.
      security:
        - bearerAuthToken: []
      parameters:
        - name: bytes
          in: query
          schema:
            type: integer
            minimum: 1
            default: 32
          description: Quantidade de bytes. Limite = MAX_BYTES_PER_REQUEST (padrão 1 MiB).
      responses:
        '200':
          description: N bytes brutos (application/octet-stream, Content-Length = N).
          headers:
            X-Request-Id:
              schema:
                type: string
              description: Identificador da requisição.
            X-QRNG-Source:
              schema:
                type: string
              description: Proveniência da fonte.
            X-QRNG-Conditioned:
              schema:
                type: string
                enum:
                  - 'false'
              description: Sempre 'false'.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: bytes acima de MAX_BYTES_PER_REQUEST.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: bytes inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          description: Upstream com erro ou formato inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Upstream indisponível ou entropia insuficiente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...